requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asyncpg
attrs==24.2.0
discord.py==2.5.2
dnspython==2.7.0
feedparser==6.0.11
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
parsedatetime==2.6
propcache==0.2.0
pydantic==2.10.6
python-dateutil==2.9.0.post0
six==1.16.0
tabulate==0.9.0
yarl==1.17.1